This commit was manufactured by cvs2svn to create tag 'r1-2M12'.
git-svn-id: https://svn.apache.org/repos/asf/incubator/cayenne/main/tags/r1-2M12@404697 13f79535-47bb-0310-9956-ffa450edef68

+++ cayenne/cayenne-java/src/cayenne/java/org/objectstyle/cayenne/util/MemoryBlob.java
@@ -246,7 +246,7 @@ public class MemoryBlob implements Blob {
 
         final long plen = pattern.length();
 
-        if (plen == 0 || start > ((long) dlen) - plen) {
+        if (plen == 0 || start > dlen - plen) {
             return -1;
         }
